# faster tree transversal