--- a/lib/compress.js
+++ b/lib/compress.js
@@ -1911,7 +1911,7 @@ merge(Compressor.prototype, {
         // returned if nothing changed.
         function trim(nodes, compressor, first_in_statement) {
             var ret = [], changed = false;
-            for (var i = 0, ii = nodes.length; i < ii; i++) {
+            for (var i = 0, len = nodes.length; i < len; i++) {
                 var node = nodes[i].drop_side_effect_free(compressor, first_in_statement);
                 changed |= node !== nodes[i];
                 if (node) {