Born out of #1460 - to separate these non-breaking changes from that PR.