@alexlamsl 2017-02-03
### master

```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.109s
- scope: 0.380s
- squeeze: 2.093s
- mangle: 0.047s
- generate: 0.110s


https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.228s
- scope: 0.642s
- squeeze: 3.133s
- mangle: 0.067s
- generate: 0.202s


https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.463s
- scope: 1.483s
- squeeze: 52.986s
- mangle: 0.144s
- generate: 0.537s


https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.040s
- scope: 0.104s
- squeeze: 0.484s
- mangle: 0.014s
- generate: 0.053s
```

### This PR

```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.108s
- scope: 0.285s
- squeeze: 1.908s
- mangle: 0.030s
- generate: 0.093s


https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.225s
- scope: 0.559s
- squeeze: 2.808s
- mangle: 0.063s
- generate: 0.175s


https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.479s
- scope: 1.432s
- squeeze: 50.035s
- mangle: 0.135s
- generate: 0.487s


https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.046s
- scope: 0.101s
- squeeze: 0.479s
- mangle: 0.012s
- generate: 0.034s
```