@kzc 2017-02-03
Way back in the day when I used [].sort() for Mozilla and IE I noticed one was a stable sort and one was not. I don't know if it is standard across all JS engines. For reproducibility reasons I'd recommend to keep mergeSort().