@alexlamsl 2017-02-03
@kzc both you and Travis CI has reached the same conclusion. The commit has been reverted.