@alexlamsl 2017-02-03
https://github.com/mishoo/UglifyJS2/issues/1145#issuecomment-227006941 with `-mc warnings=false,sequences=<N> --stats`

N | master | PR
----- | ----- | -----
100 | 11s | 8s
200 | 19s | 18s
500 | 48s | 44s
1000 | 93s | 87s