@alexlamsl 2017-02-03
### reduced usage of `[].reduce()` (https://github.com/mishoo/UglifyJS2/pull/1462/commits/bf7f4e7d935bf8b6e86f25e556973f595fc0f200)

```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.106s
- scope: 0.279s
- squeeze: 0.730s
- mangle: 0.051s
- generate: 0.088s


https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.221s
- scope: 0.535s
- squeeze: 1.048s
- mangle: 0.063s
- generate: 0.165s


https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.475s
- scope: 1.428s
- squeeze: 16.092s
- mangle: 0.136s
- generate: 0.454s


https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.041s
- scope: 0.098s
- squeeze: 0.192s
- mangle: 0.011s
- generate: 0.033s
```