@alexlamsl 2017-02-03
<sub>(manually applied #1443 in both cases before running `test/benchmark.js`)</sub>

### master
```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.128s
- scope: 0.296s
- squeeze: 0.817s
- mangle: 0.030s
- generate: 0.100s

SHA1: 1e8be2c77abae2e2b4c9e6e994848c10d0b3f724

https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.231s
- scope: 0.568s
- squeeze: 1.070s
- mangle: 0.067s
- generate: 0.187s

SHA1: d3da5ad3cc12ad8e78e164adaf8f0cb1483c9943

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.500s
- scope: 1.506s
- squeeze: 17.437s
- mangle: 0.140s
- generate: 0.495s

SHA1: b7f7c28f5c7a1fe4e807c34de983509f8ab9d6b8

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.075s
- scope: 0.126s
- squeeze: 0.230s
- mangle: 0.015s
- generate: 0.043s

SHA1: ac90c5749c1e7b35712c09a39c9fae4dbeae0d22

https://unpkg.com/react@15.3.2/dist/react.js
Timing information (compressed 1 files):
- parse: 0.220s
- scope: 0.535s
- squeeze: 1.218s
- mangle: 0.051s
- generate: 0.181s

SHA1: fab67a6f7747ef9d657ceb2dc17f61d13f83f3db

http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
Timing information (compressed 1 files):
- parse: 0.473s
- scope: 1.323s
- squeeze: 2.390s
- mangle: 0.144s
- generate: 0.463s

SHA1: 776fb4dcd73859227ed2ab6998b6e18073816c24

https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
Timing information (compressed 1 files):
- parse: 0.129s
- scope: 0.296s
- squeeze: 0.691s
- mangle: 0.059s
- generate: 0.112s

SHA1: 639b5fd837a2eec07b3bf09f29c98e22e7ee2ae2

https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
Timing information (compressed 1 files):
- parse: 0.263s
- scope: 0.815s
- squeeze: 1.586s
- mangle: 0.164s
- generate: 0.281s

SHA1: cee4cbca8cbfd36b4a90f0f77657fff2efd03988
```

### https://github.com/mishoo/UglifyJS2/pull/1462/commits/137c7c2dfa6b6a924a9b63e0896a7ddadce47b4a
```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.115s
- scope: 0.290s
- squeeze: 0.776s
- mangle: 0.031s
- generate: 0.092s

SHA1: 1e8be2c77abae2e2b4c9e6e994848c10d0b3f724

https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.239s
- scope: 0.560s
- squeeze: 1.052s
- mangle: 0.068s
- generate: 0.175s

SHA1: d3da5ad3cc12ad8e78e164adaf8f0cb1483c9943

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.543s
- scope: 1.536s
- squeeze: 16.312s
- mangle: 0.136s
- generate: 0.439s

SHA1: b7f7c28f5c7a1fe4e807c34de983509f8ab9d6b8

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.063s
- scope: 0.110s
- squeeze: 0.235s
- mangle: 0.012s
- generate: 0.038s

SHA1: ac90c5749c1e7b35712c09a39c9fae4dbeae0d22

https://unpkg.com/react@15.3.2/dist/react.js
Timing information (compressed 1 files):
- parse: 0.202s
- scope: 0.527s
- squeeze: 1.241s
- mangle: 0.050s
- generate: 0.159s

SHA1: fab67a6f7747ef9d657ceb2dc17f61d13f83f3db

http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
Timing information (compressed 1 files):
- parse: 0.519s
- scope: 1.331s
- squeeze: 2.266s
- mangle: 0.136s
- generate: 0.413s

SHA1: 776fb4dcd73859227ed2ab6998b6e18073816c24

https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
Timing information (compressed 1 files):
- parse: 0.127s
- scope: 0.277s
- squeeze: 0.637s
- mangle: 0.056s
- generate: 0.099s

SHA1: 639b5fd837a2eec07b3bf09f29c98e22e7ee2ae2

https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
Timing information (compressed 1 files):
- parse: 0.251s
- scope: 0.812s
- squeeze: 1.619s
- mangle: 0.160s
- generate: 0.261s

SHA1: cee4cbca8cbfd36b4a90f0f77657fff2efd03988
```