@kzc 2017-02-25
@alexlamsl This PR would certainly help 2.x timings. Any plans to rebase and merge into 2.x?