@alexlamsl 2017-02-25
@kzc thanks for the reminder. I'll rebase now and make some measurements.