@kzc 2017-02-25
Despite the recent [SHA1 collision news](https://arstechnica.com/security/2017/02/at-deaths-door-for-years-widely-used-sha1-function-is-now-dead/), if the SHA1s match master then it's safe to merge.