@alexlamsl 2017-02-25
`test/benchmark.js`

File | `master` | #1462
 --- | ---:| ---:
jquery-3.1.1.js | 0.948s | 0.852s
angular.js | 1.384s | 1.242s
math.js | 17.507s | 16.320s
bootstrap.js | 0.247s | 0.284s
react.js | 1.277s | 1.453s
ember.prod.js | 2.805s | 2.688s
lodash.js | 0.895s | 0.850s
d3.js | 1.848s | 1.906s
mandreel.js | 55.320s | 55.425s

`mandreel.js` is byte-identical, while others have identical SHA1s.