@kzc 2017-02-25
The timing regression in react and bootstrap is odd.