@kzc 2017-02-25
node 6.9.0 timings on Mac post-hoist look better than pre-hoist.  `jquery` and `lodash` are faster, and the rest are a wash. Oddly math.js is the same speed now.

**master:**
```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.160s
- scope: 0.323s
- squeeze: 0.977s
- mangle: 0.033s
- generate: 0.116s

SHA1: e2f1dfa47de9b2dcda43532bb65421aec0a096fc

https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.333s
- scope: 0.479s
- squeeze: 1.305s
- mangle: 0.051s
- generate: 0.163s

SHA1: 93ccb45f3d5b97ffa4337d8f4c60bdb5316c27c0

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.607s
- scope: 1.010s
- squeeze: 11.162s
- mangle: 0.077s
- generate: 0.370s

SHA1: db32e68609a7fc5d9919eaad6ef4824d325806a9

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.076s
- scope: 0.098s
- squeeze: 0.304s
- mangle: 0.008s
- generate: 0.037s

SHA1: 5c0753277091a57ffec44d7380c0d6902780231a

https://unpkg.com/react@15.3.2/dist/react.js
Timing information (compressed 1 files):
- parse: 0.365s
- scope: 0.476s
- squeeze: 1.331s
- mangle: 0.034s
- generate: 0.159s

SHA1: b9b4bc19faeaf5230ca77ac6e9bba98b7baeaf1d

http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
Timing information (compressed 1 files):
- parse: 0.637s
- scope: 1.078s
- squeeze: 2.727s
- mangle: 0.093s
- generate: 0.374s

SHA1: 6f722a82e5a0612e833fbf5f7ecfbeebfdc8d9f0

https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
Timing information (compressed 1 files):
- parse: 0.241s
- scope: 0.334s
- squeeze: 0.863s
- mangle: 0.039s
- generate: 0.105s

SHA1: d5d6f7196c5437d713d0fda63c454f54c5bd1190

https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
Timing information (compressed 1 files):
- parse: 0.374s
- scope: 0.710s
- squeeze: 1.777s
- mangle: 0.108s
- generate: 0.232s

SHA1: 26a57cefc940fcd7a7ce337b6a233125097362df
```

**this PR (after hoist)**:
```
https://code.jquery.com/jquery-3.1.1.js
Timing information (compressed 1 files):
- parse: 0.149s
- scope: 0.278s
- squeeze: 0.883s
- mangle: 0.025s
- generate: 0.102s

SHA1: e2f1dfa47de9b2dcda43532bb65421aec0a096fc

https://code.angularjs.org/1.6.1/angular.js
Timing information (compressed 1 files):
- parse: 0.332s
- scope: 0.468s
- squeeze: 1.340s
- mangle: 0.049s
- generate: 0.160s

SHA1: 93ccb45f3d5b97ffa4337d8f4c60bdb5316c27c0

https://cdnjs.cloudflare.com/ajax/libs/mathjs/3.9.0/math.js
Timing information (compressed 1 files):
- parse: 0.571s
- scope: 1.056s
- squeeze: 11.191s
- mangle: 0.086s
- generate: 0.354s

SHA1: db32e68609a7fc5d9919eaad6ef4824d325806a9

https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.js
Timing information (compressed 1 files):
- parse: 0.074s
- scope: 0.099s
- squeeze: 0.301s
- mangle: 0.008s
- generate: 0.037s

SHA1: 5c0753277091a57ffec44d7380c0d6902780231a

https://unpkg.com/react@15.3.2/dist/react.js
Timing information (compressed 1 files):
- parse: 0.300s
- scope: 0.534s
- squeeze: 1.313s
- mangle: 0.042s
- generate: 0.154s

SHA1: b9b4bc19faeaf5230ca77ac6e9bba98b7baeaf1d

http://builds.emberjs.com/tags/v2.11.0/ember.prod.js
Timing information (compressed 1 files):
- parse: 0.570s
- scope: 1.083s
- squeeze: 2.786s
- mangle: 0.090s
- generate: 0.358s

SHA1: 6f722a82e5a0612e833fbf5f7ecfbeebfdc8d9f0

https://cdn.jsdelivr.net/lodash/4.17.4/lodash.js
Timing information (compressed 1 files):
- parse: 0.176s
- scope: 0.256s
- squeeze: 0.773s
- mangle: 0.036s
- generate: 0.094s

SHA1: d5d6f7196c5437d713d0fda63c454f54c5bd1190

https://cdnjs.cloudflare.com/ajax/libs/d3/4.5.0/d3.js
Timing information (compressed 1 files):
- parse: 0.362s
- scope: 0.677s
- squeeze: 1.707s
- mangle: 0.111s
- generate: 0.246s

SHA1: 26a57cefc940fcd7a7ce337b6a233125097362df
```